Move Prop/Jet/Heli class effects to Aircraft_base.py
